Added Google OAuth authentication for supervisors and admins.
Removed the old username and password based authentication and added Google OAuth. Now, only the supervisors and admins are able to login. Note that, still the data filtering for different users are not complete, hence all the data are shown. Also requires to manually create supervisor and admin roles in database.